Fix for issue 6
Warfarin appears in the search results even though the page for it is
just a list of indications, and this raises an error when the app
tries to render it.  Until another template is added for indication
lists (if that's what they're call, if not they should be) this fix
will stop drugs without a name key appearing in search results.